bugfix + draft version sending data to kafka after saving it in mongo
